Installs ts-node, nodemon
These two tools, ts-node[1] and nodemon[2], will make local development
way easier. nodemon will restart the server anytime a file changes
(using ts-node to recomplie typescript files on the fly) so that we're
not constantly restarting the server manually.

Also, I'm introducing the dev script; this is separate from start
because start is still what we want to use in a production environment.

[1] https://typestrong.org/ts-node/docs/
[2] https://nodemon.io/