Adds ClueId
This type represents the identifier of a clue *within a single puzzle*,
e.g. "4a" or "12d". This is not a universally unique identifier! Many
clues will have the same one.

We also include two functions, fromString and toString, for converting
to and from this type. It's going to be common for us to accept clue ids
as user input, so it will be handy to have a way to safely refine that
flexible input down to a well known type.