Introduces Clue Type with findRandomById
The Clue type models a clue from George Ho's dataset (but a bit cleaned
up and easier to work with).

findRandomById is a function that takes a ClueId (reminder: not
universally unique, just unique within a puzzle) and returns a random
clue that has that id; e.g. "2a", "21d", etc.